+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/workflows/dispatcher/HandlersFactory.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Hedera Hashgraph, LLC
+ * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -87,7 +87,7 @@ public final class HandlersFactory {
         requireNonNull(hederaAccountNumbers);
         requireNonNull(hederaFileNumbers);
 
-        // TODO: Add required parameters to constructors
+        // TODO - provide DI, c.f. https://github.com/hashgraph/hedera-services/issues/4317
         return new Handlers(
                 new ConsensusCreateTopicHandler(),
                 new ConsensusUpdateTopicHandler(),
