--- hedera-node/hedera-token-service-impl/src/main/java/com/hedera/node/app/service/token/impl/CryptoPreTransactionHandlerImpl.java
@@ -1,206 +1,0 @@
-/*
- * Copyright (C) 2022 Hedera Hashgraph, LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hedera.node.app.service.token.impl;
-
-import static com.hedera.node.app.service.mono.Utils.asHederaKey;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_ALLOWANCE_OWNER_ID;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_DELEGATING_SPENDER;
-import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_TRANSFER_ACCOUNT_ID;
-
-import com.google.common.annotations.VisibleForTesting;
-import com.hedera.node.app.service.token.CryptoPreTransactionHandler;
-import com.hedera.node.app.spi.PreHandleContext;
-import com.hedera.node.app.spi.key.HederaKey;
-import com.hedera.node.app.spi.meta.SigTransactionMetadataBuilder;
-import com.hedera.node.app.spi.meta.TransactionMetadata;
-import com.hederahashgraph.api.proto.java.AccountID;
-import com.hederahashgraph.api.proto.java.TransactionBody;
-import edu.umd.cs.findbugs.annotations.NonNull;
-import java.util.Objects;
-import java.util.Optional;
-import org.apache.commons.lang3.NotImplementedException;
-
-/**
- * A {@code CryptoPreTransactionHandler} implementation that pre-computes the required signing keys
- * (but not the candidate signatures) for each crypto operation.
- */
-public final class CryptoPreTransactionHandlerImpl implements CryptoPreTransactionHandler {
-    private final ReadableAccountStore accountStore;
-    private final PreHandleContext preHandleContext;
-    private CryptoSignatureWaiversImpl waivers;
-
-    public CryptoPreTransactionHandlerImpl(
-            @NonNull final ReadableAccountStore accountStore, @NonNull final PreHandleContext ctx) {
-        this.accountStore = Objects.requireNonNull(accountStore);
-        this.preHandleContext = Objects.requireNonNull(ctx);
-        this.waivers = new CryptoSignatureWaiversImpl(preHandleContext.accountNumbers());
-    }
-
-    @Override
-    /** {@inheritDoc} */
-    public TransactionMetadata preHandleCryptoCreate(
-            final TransactionBody tx, final AccountID payer) {
-        final var op = tx.getCryptoCreateAccount();
-        final var key = asHederaKey(op.getKey());
-        final var receiverSigReq = op.getReceiverSigRequired();
-        return createAccountSigningMetadata(tx, key, receiverSigReq, payer);
-    }
-
-    @Override
-    /** {@inheritDoc} */
-    public TransactionMetadata preHandleCryptoDelete(
-            final TransactionBody txn, final AccountID payer) {
-        final var op = txn.getCryptoDelete();
-        final var deleteAccountId = op.getDeleteAccountID();
-        final var transferAccountId = op.getTransferAccountID();
-        final var meta =
-                new SigTransactionMetadataBuilder(accountStore)
-                        .payerKeyFor(payer)
-                        .txnBody(txn)
-                        .addNonPayerKey(deleteAccountId)
-                        .addNonPayerKeyIfReceiverSigRequired(
-                                transferAccountId, INVALID_TRANSFER_ACCOUNT_ID);
-        return meta.build();
-    }
-
-    @Override
-    /** {@inheritDoc} */
-    public TransactionMetadata preHandleApproveAllowances(
-            final TransactionBody txn, final AccountID payer) {
-        final var op = txn.getCryptoApproveAllowance();
-        final var meta =
-                new SigTransactionMetadataBuilder(accountStore).payerKeyFor(payer).txnBody(txn);
-        var failureStatus = INVALID_ALLOWANCE_OWNER_ID;
-
-        for (final var allowance : op.getCryptoAllowancesList()) {
-            meta.addNonPayerKey(allowance.getOwner(), failureStatus);
-        }
-        for (final var allowance : op.getTokenAllowancesList()) {
-            meta.addNonPayerKey(allowance.getOwner(), failureStatus);
-        }
-        for (final var allowance : op.getNftAllowancesList()) {
-            final var ownerId = allowance.getOwner();
-            // If a spender who is granted approveForAll from owner and is granting
-            // allowance for a serial to another spender, need signature from the approveForAll
-            // spender
-            var operatorId =
-                    allowance.hasDelegatingSpender() ? allowance.getDelegatingSpender() : ownerId;
-            // If approveForAll is set to true, need signature from owner
-            // since only the owner can grant approveForAll
-            if (allowance.getApprovedForAll().getValue()) {
-                operatorId = ownerId;
-            }
-            if (operatorId != ownerId) {
-                failureStatus = INVALID_DELEGATING_SPENDER;
-            }
-            meta.addNonPayerKey(operatorId, failureStatus);
-        }
-        return meta.build();
-    }
-
-    @Override
-    /** {@inheritDoc} */
-    public TransactionMetadata preHandleDeleteAllowances(
-            final TransactionBody txn, final AccountID payer) {
-        final var op = txn.getCryptoDeleteAllowance();
-        final var meta =
-                new SigTransactionMetadataBuilder(accountStore).payerKeyFor(payer).txnBody(txn);
-        // Every owner whose allowances are being removed should sign, if the owner is not payer
-        for (final var allowance : op.getNftAllowancesList()) {
-            meta.addNonPayerKey(allowance.getOwner(), INVALID_ALLOWANCE_OWNER_ID);
-        }
-        return meta.build();
-    }
-
-    @Override
-    /** {@inheritDoc} */
-    public TransactionMetadata preHandleUpdateAccount(
-            final TransactionBody txn, final AccountID payer) {
-        final var op = txn.getCryptoUpdateAccount();
-        final var updateAccountId = op.getAccountIDToUpdate();
-        final var meta =
-                new SigTransactionMetadataBuilder(accountStore).payerKeyFor(payer).txnBody(txn);
-
-        final var newAccountKeyMustSign = !waivers.isNewKeySignatureWaived(txn, payer);
-        final var targetAccountKeyMustSign = !waivers.isTargetAccountSignatureWaived(txn, payer);
-        if (targetAccountKeyMustSign) {
-            meta.addNonPayerKey(updateAccountId);
-        }
-        if (newAccountKeyMustSign && op.hasKey()) {
-            final var candidate = asHederaKey(op.getKey());
-            candidate.ifPresent(meta::addToReqNonPayerKeys);
-        }
-        return meta.build();
-    }
-
-    @Override
-    /** {@inheritDoc} */
-    public TransactionMetadata preHandleCryptoTransfer(
-            final TransactionBody txn, final AccountID payer) {
-        throw new NotImplementedException();
-    }
-
-    @Override
-    /** {@inheritDoc} */
-    public TransactionMetadata preHandleAddLiveHash(
-            final TransactionBody txn, final AccountID payer) {
-        throw new NotImplementedException();
-    }
-
-    @Override
-    /** {@inheritDoc} */
-    public TransactionMetadata preHandleDeleteLiveHash(
-            final TransactionBody txn, final AccountID payer) {
-        throw new NotImplementedException();
-    }
-
-    /* --------------- Helper methods --------------- */
-
-    /**
-     * Returns metadata for {@code CryptoCreate} transaction needed to validate signatures needed
-     * for signing the transaction
-     *
-     * @param txn given transaction body
-     * @param key key provided in the transaction body
-     * @param receiverSigReq flag for receiverSigReq on the given transaction body
-     * @param payer payer for the transaction
-     * @return transaction's metadata needed to validate signatures
-     */
-    private TransactionMetadata createAccountSigningMetadata(
-            final TransactionBody txn,
-            final Optional<HederaKey> key,
-            final boolean receiverSigReq,
-            final AccountID payer) {
-        final var meta =
-                new SigTransactionMetadataBuilder(accountStore).payerKeyFor(payer).txnBody(txn);
-        if (receiverSigReq && key.isPresent()) {
-            meta.addToReqNonPayerKeys(key.get());
-        }
-        return meta.build();
-    }
-
-    /**
-     * @param waivers signature waivers for crypto service
-     * @deprecated This method is needed for testing until {@link CryptoSignatureWaiversImpl} is
-     *     implemented. FUTURE: This method should be removed once {@link
-     *     CryptoSignatureWaiversImpl} is implemented.
-     */
-    @Deprecated(forRemoval = true)
-    @VisibleForTesting
-    public void setWaivers(final CryptoSignatureWaiversImpl waivers) {
-        this.waivers = waivers;
-    }
-}
