+++ hedera-node/hedera-token-service-impl/src/main/java/com/hedera/node/app/service/token/impl/handlers/CryptoApproveAllowanceHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 Hedera Hashgraph, LLC
+ * Copyright (C) 2022-2023 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,11 @@
  */
 package com.hedera.node.app.service.token.impl.handlers;
 
+import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_ALLOWANCE_OWNER_ID;
+import static com.hederahashgraph.api.proto.java.ResponseCodeEnum.INVALID_DELEGATING_SPENDER;
+
 import com.hedera.node.app.service.token.impl.ReadableAccountStore;
+import com.hedera.node.app.spi.meta.SigTransactionMetadataBuilder;
 import com.hedera.node.app.spi.meta.TransactionMetadata;
 import com.hedera.node.app.spi.workflows.TransactionHandler;
 import com.hederahashgraph.api.proto.java.AccountID;
@@ -27,7 +31,6 @@ import edu.umd.cs.findbugs.annotations.NonNull;
  * com.hederahashgraph.api.proto.java.HederaFunctionality#CryptoApproveAllowance}.
  */
 public class CryptoApproveAllowanceHandler implements TransactionHandler {
-
     /**
      * This method is called during the pre-handle workflow.
      *
@@ -38,7 +41,7 @@ public class CryptoApproveAllowanceHandler implements TransactionHandler {
      * <p>Please note: the method signature is just a placeholder which is most likely going to
      * change.
      *
-     * @param txBody the {@link TransactionBody} with the transaction data
+     * @param txn the {@link TransactionBody} with the transaction data
      * @param payer the {@link AccountID} of the payer
      * @param accountStore the {@link ReadableAccountStore} with the current data
      * @return the {@link TransactionMetadata} with all information that needs to be passed to
@@ -46,10 +49,38 @@ public class CryptoApproveAllowanceHandler implements TransactionHandler {
      * @throws NullPointerException if one of the arguments is {@code null}
      */
     public TransactionMetadata preHandle(
-            @NonNull final TransactionBody txBody,
+            @NonNull final TransactionBody txn,
             @NonNull final AccountID payer,
             @NonNull final ReadableAccountStore accountStore) {
-        throw new UnsupportedOperationException("Not implemented");
+        final var op = txn.getCryptoApproveAllowance();
+        final var meta =
+                new SigTransactionMetadataBuilder(accountStore).payerKeyFor(payer).txnBody(txn);
+        var failureStatus = INVALID_ALLOWANCE_OWNER_ID;
+
+        for (final var allowance : op.getCryptoAllowancesList()) {
+            meta.addNonPayerKey(allowance.getOwner(), failureStatus);
+        }
+        for (final var allowance : op.getTokenAllowancesList()) {
+            meta.addNonPayerKey(allowance.getOwner(), failureStatus);
+        }
+        for (final var allowance : op.getNftAllowancesList()) {
+            final var ownerId = allowance.getOwner();
+            // If a spender who is granted approveForAll from owner and is granting
+            // allowance for a serial to another spender, need signature from the approveForAll
+            // spender
+            var operatorId =
+                    allowance.hasDelegatingSpender() ? allowance.getDelegatingSpender() : ownerId;
+            // If approveForAll is set to true, need signature from owner
+            // since only the owner can grant approveForAll
+            if (allowance.getApprovedForAll().getValue()) {
+                operatorId = ownerId;
+            }
+            if (operatorId != ownerId) {
+                failureStatus = INVALID_DELEGATING_SPENDER;
+            }
+            meta.addNonPayerKey(operatorId, failureStatus);
+        }
+        return meta.build();
     }
 
     /**
