+++ hedera-node/hedera-token-service-impl/src/test/java/module-info.java
@@ -18,4 +18,6 @@ module com.hedera.node.app.service.token.impl.test {
             org.junit.platform.commons;
     opens com.hedera.node.app.service.token.impl.test.util to
             org.junit.platform.commons;
+    opens com.hedera.node.app.service.token.impl.test.handlers to
+            org.junit.platform.commons;
 }
